Changed setMotorVelocity to setMotorPower

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/controllers/TeleOpController.java
@@ -45,7 +45,7 @@ public class TeleOpController {
      * @param fieldCentric Whether to use field centric driving
      * @param usePedro Whether to use Pedro Pathing
      */
-    public void drivetrainLogic(boolean fieldCentric, boolean usePedro, int speed) {
+    public void drivetrainLogic(boolean fieldCentric, boolean usePedro, double speed) {
         if (gamepad1.dpadLeftWasPressed()) robotCentric = true;
         else if (gamepad1.dpadRightWasPressed()) robotCentric = false;
         if (gamepad1.leftBumperWasPressed() && fred.doesClawsExist()) fred.toggleClaws();
